backend/requirements.txt
fastapi
uvicorn

pyJWT
python-decouple

pymongo